Covariance estimation
Fixed covariance estimation. Now the implementation produces an (almost) equivalent p-value to the current version of dearseq when covariance structures are not estimated. 

Found a bug wherein if gene names were duplicated in the gene set, the gene would be removed from the expression matrix. Added a case to catch this and make sure the gene names within every gene set are unique - this is why the current version of dearseq gives a slightly different p-value.

Implementation is still less efficient than current dearseq even without estimating the correlation structures.